refactor: simplify runtime attachment retry
