+++ packages/frontend/src/pages/agents/use-agent-studio-runtime-attachment-retry.test.ts
@@ -96,6 +96,26 @@ describe("selectRuntimeAttachmentCandidates", () => {
     expect(withEquivalentPathFormatting).toEqual(baseline);
   });
 
+  test("deduplicates duplicate runtime summaries", () => {
+    const candidates = selectRuntimeAttachmentCandidates({
+      repoPath: "/repo",
+      session: {
+        runtimeKind: "opencode",
+      },
+      runtimeSources: [
+        { kind: "opencode", repoPath: "/repo" },
+        { kind: "opencode", repoPath: "/repo/" },
+      ],
+    });
+
+    expect(candidates).toEqual([
+      {
+        runtimeKind: "opencode",
+        repoPath: "/repo",
+      },
+    ]);
+  });
+
   test("compares candidates structurally", () => {
     expect(
       haveSameRuntimeAttachmentCandidates(
